move `IntoFileName` trait to live alongside `FileName`
also, use it.
