+++ src/build.rs
@@ -5,12 +5,12 @@ use languageserver_types::Position;
 use lark_entity::{EntityData, ItemKind, MemberKind};
 use lark_language_server::{lsp_serve, LspResponder};
 use lark_mir::MirDatabase;
-use lark_parser::{IntoFileName, ParserDatabase, ParserDatabaseExt};
+use lark_parser::{ParserDatabase, ParserDatabaseExt};
 use lark_query_system::ls_ops::Cancelled;
 use lark_query_system::ls_ops::LsDatabase;
 use lark_query_system::LarkDatabase;
 use lark_query_system::QuerySystem;
-use lark_span::{ByteIndex, FileName, Span};
+use lark_span::{ByteIndex, FileName, IntoFileName, Span};
 use lark_task_manager::Actor;
 use salsa::Database;
 use std::borrow::Cow;
